Move init tests (#7318)
